+++ classes/external.php
@@ -367,6 +367,7 @@ class quizaccess_proctoring_external extends external_api {
             'face_missing',
             'no_face_detected',
             'audio_detected',
+            'phone_detected',
             'multiple_monitors_detected',
             'monitor_detection_unavailable',
         ];